+++ src/components/core/inputs/index.ts
@@ -1,3 +1,4 @@
 export { default as Checkbox } from './Checkbox';
+export { default as Dropdown } from './Dropdown';
 export { default as Form } from './Form';
 export { default as TextInput } from './TextInput';